facade dwt function
Added a function to decompose the signal into one level of approximation
and detail coeffs. Added utility methods like pad, upsample etc.
Currently the results of dwt are *not* same as MATLAB dwt because MATLAB
dwt does a lot of padding for the reason I need to figure out. The
results do match with MATLAB downsample(conv(sig, filter)) output.

TODO: Figure out the problem with the sign reversal of high pass filter
result. Figure out padding. Finish iDwt so that we can move on to cascading